requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
langgraph
langchain

langchain-core

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai